# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/user/register":{"post":{"operationId":"register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}},"/user/{id}":{"get":{"operationId":"getOne","summary":"Retrieve a single User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"fields","description":"Selects resource fields. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#select\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":false},{"name":"join","description":"Adds relational resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#join\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"cache","description":"Reset cache (if was enabled). <a href=\"https://github.com/nestjsx/crud/wiki/Requests#cache\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer","minimum":0,"maximum":1}}],"responses":{"200":{"description":"Get one base response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"patch":{"operationId":"updateOne","summary":"Update a single User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"put":{"operationId":"replaceOne","summary":"Replace a single User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"delete":{"operationId":"deleteOne","summary":"Delete a single User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Delete one base response"}},"tags":["User"]}},"/user":{"get":{"operationId":"getMany","summary":"Retrieve multiple Users","parameters":[{"name":"fields","description":"Selects resource fields. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#select\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":false},{"name":"s","description":"Adds search condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#search\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"string"}},{"name":"filter","description":"Adds filter condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#filter\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"or","description":"Adds OR condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#or\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"sort","description":"Adds sort by field. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#sort\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"join","description":"Adds relational resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#join\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"limit","description":"Limit amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#limit\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"offset","description":"Offset amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#offset\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"page","description":"Page portion of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#page\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"cache","description":"Reset cache (if was enabled). <a href=\"https://github.com/nestjsx/crud/wiki/Requests#cache\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer","minimum":0,"maximum":1}}],"responses":{"200":{"description":"Get many base response","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/GetManyUserResponseDto"},{"type":"array","items":{"$ref":"#/components/schemas/User"}}]}}}}},"tags":["User"]},"post":{"operationId":"createOne","summary":"Create a single User","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Get create one base response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}},"/user/bulk":{"post":{"operationId":"createMany","summary":"Create multiple Users","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateManyUserDto"}}}},"responses":{"201":{"description":"Get create many base response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"]}},"/auth/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/test":{"get":{"operationId":"test","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/role":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["role"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["role"]}},"/role/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["role"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["role"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["role"]}},"/faster-crud/dict/{entity}":{"get":{"operationId":"getDict","parameters":[{"name":"entity","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["fce"]}},"/faster-crud/all":{"get":{"operationId":"getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["fce"]}},"/file/upload":{"post":{"operationId":"uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["File"]}},"/file/{key}":{"get":{"operationId":"getFile","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["File"]}}},"info":{"title":"sanpu API","description":"Ai Song","version":"1.0","contact":{}},"tags":[{"name":"aisong","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","minLength":1,"maxLength":20},"password":{"type":"string","minLength":1,"maxLength":32}},"required":["username","password"]},"GetManyUserResponseDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"count":{"type":"number"},"total":{"type":"number"},"page":{"type":"number"},"pageCount":{"type":"number"}},"required":["data","count","total","page","pageCount"]},"Role":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","name","user"]},"User":{"type":"object","properties":{"id":{"type":"number"},"username":{"type":"string"},"password":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Role"}}},"required":["id","username","password","roles"]},"CreateManyUserDto":{"type":"object","properties":{"bulk":{"type":"array","items":{"$ref":"#/components/schemas/User"}}},"required":["bulk"]},"PasswordLoginDto":{"type":"object","properties":{"username":{"type":"string","minLength":1,"maxLength":20},"password":{"type":"string","minLength":1,"maxLength":32}},"required":["username","password"]},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}}}}}