# src/modules/role/entities/role.entity.ts
import { User } from '@/modules/user/entities/user.entity'
import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm'

@Entity()
export class Role {
  @PrimaryGeneratedColumn()
  id: number

  @Column()
  name: string

  @ManyToOne(() => User, (user) => user.roles)
  user: User

  constructor(name: string) {
    this.name = name
  }
}